registro documento+++

+++ src/app/dialogs/dialog-delete/alert-dialog-delete.component.ts
@@ -3,6 +3,7 @@ import {MAT_DIALOG_DATA, MatDialog, MatDialogRef} from '@angular/material/dialog
 import {Observable, Observer} from 'rxjs';
 import {WebServiceAPIService} from '../../api/web-service-api.service';
 import {Section} from '../../admin/actividades/index/index.component';
+import {FileData} from '../../client/registrar-documento/registrar-documento.component';
 
 export class DataModal {
   constructor(public title: string, public message: string, public typeObject: string, public  folder: Section) {
@@ -20,7 +21,7 @@ export class AlertDialogDelete implements OnInit {
 
   title = '';
   message = '';
-  section: Section;
+  objectItem: any;
   typeObject = '';
   public itemDeleted = false;
   isHideProggres = true;
@@ -32,7 +33,7 @@ export class AlertDialogDelete implements OnInit {
   ) {
     this.title = data.title;
     this.message = data.message;
-    this.section = data.folder;
+    this.objectItem = data.folder;
     this.typeObject = data.typeObject;
   }
 
@@ -44,19 +45,29 @@ export class AlertDialogDelete implements OnInit {
   }
 
   elimnar() {
-    let idItem = this.section.id;
-    let dialog = this.dialogRef;
     let isHideProggresLocal = false;
+    if (this.typeObject == 'FILE') {
 
-    this.isHideProggres = isHideProggresLocal;
-    this.service.deleteItem(this.typeObject, idItem, dialog).then(function() {
-      console.log('Document successfully deleted!');
-      dialog.close('Deleted');
-      isHideProggresLocal = true;
-    }).catch(function(error) {
-      console.error('Error removing document: ', error);
-      isHideProggresLocal = true;
-    });
+      let itemFile: FileData = this.objectItem;
+      this.dialogRef.close('F_DEL');
+
+    } else {
+      let idItem = this.objectItem.id;
+      let dialog = this.dialogRef;
+
+
+      this.isHideProggres = isHideProggresLocal;
+      this.service.deleteItem(this.typeObject, idItem, dialog).then(function() {
+        console.log('Document successfully deleted!');
+        dialog.close('Deleted');
+        isHideProggresLocal = true;
+      }).catch(function(error) {
+        console.error('Error removing document: ', error);
+        isHideProggresLocal = true;
+      });
+    }
+
+
   }
 }
